+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 #### 功能构建
 
+- 实现 notify 模块的路由器、调度器、发送器。
+  - com.dwarfeng.familyhelper.plugin.notify.handler.dispatcher.GeneralDispatcherRegistry。
+  - com.dwarfeng.familyhelper.plugin.notify.handler.router.PermissionRouterRegistry。
+  - com.dwarfeng.familyhelper.plugin.notify.handler.sender.BuiltinSenderRegistry。
+
 - 项目建立，清除测试通过。
 
 #### Bug修复